+++ CHANGELOG.md
@@ -9,8 +9,15 @@ Instead, edit the files in dev/changes/, then run 'make docs' to update this fil
 ### Fixed
 
 ### Added
+- New input `product_version_meta` for adding a metadata field to the product version.
+- New input `work_dir` for setting the working directory.
 
 ### Changed
+- The `product_version` input is now optional. Instead we should use a VERSION file
+  in either the repo root, or .release/ or dev/.
+- The `product_name` input's default is now the working directory name, unless
+  the working directory is the repository root, in which case it still defaults
+  to the repository name.
 
 ---
 Internal Changes
@@ -19,6 +26,7 @@ Internal Changes
   - When pre-push hook fails you now see the output so it's easier to debug.
   - Readme tidied up.
   - Converted digest inputs to Go.
+  - Converted build functionality to Go.
 
 ## [v0.1.3](https://github.com/hashicorp/actions-go-build/releases/tag/v0.1.3) - June 15, 2022